(MODULES-3133) Update expected line endings for UTF8
This commit modifies the default line ending for UTF-8 files with a BOM
from a UTF-16LE CRLF to a UTF-8 LF.  This was causing some MOF files
to fail parsing do to inappropriate chomping of the line terminator.  Also
a warning was added to UTF16 files that these files are expected to have
CRLF line endings.